// cp2_6.15_auto64/_solo_repo/cp1_6.13_auto14/src/types.ts
export interface VoteOption {
  id: string;
  text: string;
  color: string;
  votes: number;
}

export interface Poll {
  _id?: string;
  title: string;
  options: VoteOption[];
  duration: number;
  status: 'pending' | 'active' | 'closed';
  createdAt: number;
  startedAt?: number;
  closedAt?: number;
}

export interface Comment {
  _id?: string;
  pollId: string;
  text: string;
  createdAt: number;
}

export type Role = 'host' | 'viewer';